# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AeroTwin API
  version: 1.0.0
paths:
...
  /api/v1/engines:
    get:
      tags:
      - engines
      summary: List Engines
      description: List all engines. Any authenticated user.
      operationId: list_engines_api_v1_engines_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Engine'
                type: array
                title: Response List Engines Api V1 Engines Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - engines
      summary: Create Engine
      description: Create a new engine. Admin or Program Manager only.
      operationId: create_engine_api_v1_engines_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Engine'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/engines/{engine_id}:
    get:
      tags:
      - engines
      summary: Get Engine
      description: Get a single engine by ID.
      operationId: get_engine_api_v1_engines__engine_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: engine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Engine'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - engines
      summary: Update Engine
      description: Update an engine. Admin, Program Manager, or Maintenance Engineer.
      operationId: update_engine_api_v1_engines__engine_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: engine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engine Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Engine'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - engines
      summary: Delete Engine
      description: Delete an engine. Admin only.
      operationId: delete_engine_api_v1_engines__engine_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: engine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engine Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/engines/{engine_id}/telemetry/latest:
    get:
      tags:
      - engines
      summary: Get Latest Telemetry
      description: Get the most recent telemetry reading for an engine.
      operationId: get_latest_telemetry_api_v1_engines__engine_id__telemetry_latest_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: engine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryReading'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engines/{engine_id}/rul:
    get:
      tags:
      - engines
      summary: Get Rul Predictions
      description: 'Get the most recent RUL predictions for an engine, optionally ranged

        by `from`/`to`. NOTE: this endpoint (not a separate `api/v1/rul.py`

        router) is the one and only implementation of `GET .../rul` — a

        second router mounted under the same `engines` prefix at the same

        path would create an ambiguous route, so `rul.py` is intentionally

        left empty. See the Phase 4.0 plan notes.'
      operationId: get_rul_predictions_api_v1_engines__engine_id__rul_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: engine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engine Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RulPrediction'
                title: Response Get Rul Predictions Api V1 Engines  Engine Id  Rul Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engines/{engine_id}/rul/status:
    get:
      tags:
      - engines
      summary: Get Rul Status
      description: 'Accuracy-First Phase 2. Distinguishes "this engine has never had

        enough telemetry to produce a RUL prediction yet" from "no

        prediction has ever been persisted" (previously indistinguishable —

        both showed as an empty `GET /rul` list). Reads `rul_service`''s

        in-memory window depth directly — no DB query, O(1), always

        up to date with the live ingestion process this API server runs in.'
      operationId: get_rul_status_api_v1_engines__engine_id__rul_status_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: engine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engines/{engine_id}/physics-consistency:
    get:
      tags:
      - engines
      summary: Get Physics Consistency
      description: 'Accuracy-First Phase 4. The latest expected-vs-measured comparison

        for each of the 5 channels the Otto-cycle physics model covers

        (CHT, EGT, oil pressure, oil temp, fuel flow) — computed by every

        live ingestion cycle (`physics_ingestion.py`) and persisted to

        `physics_deviation_readings`, one row per (engine, ts, parameter).

        Uses Postgres `DISTINCT ON` since each channel can in principle be

        updated at a slightly different cadence — this always returns each

        channel''s own most recent row, not just the 5 rows from one ts.'
      operationId: get_physics_consistency_api_v1_engines__engine_id__physics_consistency_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: engine_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Engine:
      properties:
        serial_number:
          type: string
          title: Serial Number
        uav_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Uav Asset Id
        status:
          type: string
          title: Status
          default: operational
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - serial_number
      - id
      title: Engine
    EngineCreate:
      properties:
        serial_number:
          type: string
          title: Serial Number
        uav_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Uav Asset Id
        status:
          type: string
          title: Status
          default: operational
      type: object
      required:
      - serial_number
      title: EngineCreate
    EngineUpdate:
      properties:
        serial_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Serial Number
        uav_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Uav Asset Id
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: EngineUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RulPrediction:
      properties:
        engine_id:
          type: string
          format: uuid
          title: Engine Id
        ts:
          type: string
          format: date-time
          title: Ts
        rul_cycles:
          type: number
          title: Rul Cycles
        degradation_index:
          type: number
          title: Degradation Index
        rul_lower:
          anyOf:
          - type: number
          - type: 'null'
          title: Rul Lower
        rul_upper:
          anyOf:
          - type: number
          - type: 'null'
          title: Rul Upper
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      required:
      - engine_id
      - ts
      - rul_cycles
      - degradation_index
      title: RulPrediction
...
    TelemetryReading:
      properties:
        engine_id:
          type: string
          format: uuid
          title: Engine Id
        mission_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Mission Id
        ts:
          type: string
          format: date-time
          title: Ts
        rpm:
          type: number
          title: Rpm
        cht:
          type: number
          title: Cht
        egt:
          type: number
          title: Egt
        oil_pressure:
          type: number
          title: Oil Pressure
        oil_temp:
          type: number
          title: Oil Temp
        fuel_flow:
          type: number
          title: Fuel Flow
        vibration_x:
          anyOf:
          - type: number
          - type: 'null'
          title: Vibration X
        vibration_y:
          anyOf:
          - type: number
          - type: 'null'
          title: Vibration Y
        vibration_z:
          anyOf:
          - type: number
          - type: 'null'
          title: Vibration Z
        throttle:
          anyOf:
          - type: number
          - type: 'null'
          title: Throttle
        altitude_m:
          anyOf:
          - type: number
          - type: 'null'
          title: Altitude M
        quality_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Quality Status
      type: object
      required:
      - engine_id
      - ts
      - rpm
      - cht
      - egt
      - oil_pressure
      - oil_temp
      - fuel_flow
      title: TelemetryReading
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login